[UPD] improve findingcontent component, strip off logic from form component and add it to reportForm, update function component
