+++ src/utils/submitFindingsUtils/functions.js
@@ -1,91 +1,17 @@
 import { riskField } from "./fields";
 import { initialState } from "./state";
-const submissionUrl = `/.netlify/functions/submit-finding`;
 
-const config = {
+export const config = {
   labelAll: "bug",
 };
 
-const checkTitle = (title, risk) => {
+export const checkTitle = (title, risk) => {
   if (risk === "G (Gas Optimization)") {
     return "Gas Optimizations";
   } else if (risk === "QA (Quality Assurance)") {
     return "QA Report";
   } else {
     return title;
-  }
-};
-
-export const wardenField = (wardens) => {
-  return {
-    name: "handle",
-    label: "Handle",
-    helpText: "Handle you're competing under (individual or team name)",
-    widget: "warden",
-    required: true,
-    options: wardens,
-  };
-};
-
-export const updateLocalStorage = (state, contest) => {
-  if (typeof window !== `undefined`) {
-    window.localStorage.setItem(contest, JSON.stringify(state));
-  }
-};
-
-export const handleSubmit = (
-  contest,
-  state,
-  isQaOrGasFinding,
-  setHasValidationErrors,
-  submitFinding,
-  setIsExpanded
-) => {
-  const locString = state.linesOfCode.map((loc) => loc.value).join("\n");
-  const details = isQaOrGasFinding ? state.qaGasDetails : state.details;
-  const markdownBody = `# Lines of code\n\n${locString}\n\n\n# Vulnerability details\n\n${details}\n\n`;
-
-  // extract required fields from field data for validation check
-  const formatedRisk = state.risk ? state.risk.slice(0, 1) : "";
-  const formatedBody = isQaOrGasFinding ? details : markdownBody;
-  // console.log("state", state);
-  // console.log("details", details);
-  // console.log("gas ?", isQaOrGasFinding);
-  // console.log("markdown", markdownBody);
-  const { email, handle, address, title } = state;
-  const requiredFields = isQaOrGasFinding
-    ? [email, handle, address, formatedRisk, formatedBody]
-    : [email, handle, address, formatedRisk, title, formatedBody];
-  let hasErrors = requiredFields.some((field) => {
-    return field === "" || field === undefined;
-  });
-
-  // TODO: verify that loc include code lines and are valid URLs
-  if (!isQaOrGasFinding && !state.linesOfCode[0].value) {
-    hasErrors = true;
-  }
-
-  const regex = new RegExp("#L", "g");
-  const hasInvalidLinks = state.linesOfCode.some((line) => {
-    return !regex.test(line.value);
-  });
-
-  setHasValidationErrors(hasErrors || hasInvalidLinks);
-  if (!hasErrors) {
-    submitFinding(submissionUrl, {
-      ...state,
-      body: formatedBody,
-      title: checkTitle(state.title, state.risk),
-      risk: formatedRisk
-    });
-    if (typeof window !== `undefined`) {
-      window.localStorage.removeItem(contest);
-    }
-    // New in the object passed :
-    // - LOC
-    // - details
-    // - qaGasDetails
-    setIsExpanded(false);
   }
 };
 
@@ -99,11 +25,10 @@ export const initStateFromStorage = (contest, sponsor, repoUrl, setState) => {
         (element) => element.value === dataObject.risk
       );
     }
-
     setState({
       contest: contest,
       sponsor: sponsor,
-      repo: repoUrl.split("/").pop(),
+      repo: repoUrl?.split("/").pop(),
       labels: dataObject?.labels || [config.labelAll, ""],
       title: dataObject?.title || "",
       email: dataObject?.email || "",
@@ -125,43 +50,10 @@ export const initStateFromStorage = (contest, sponsor, repoUrl, setState) => {
   }
 };
 
-export const changeHandler = (setState, e) => {
-  if (Array.isArray(e)) {
-    setState((state) => {
-      return { ...state, linesOfCode: e };
-    });
-  } else {
-    const { name, value } = e?.target;
-    switch (name) {
-      case "risk":
-        // const riskLevel = value.slice(0, 1);
-        // if (riskLevel === "G" || riskLevel === "Q") {
-        //   setIsQaOrGasFinding(true);
-        // } else {
-        //   setIsQaOrGasFinding(false);
-        // }
-        setState((state) => {
-          return {
-            ...state,
-            [name]: value,
-            labels: [config.labelAll, value ? value : ""],
-          };
-        });
-        break;
-      case "title":
-        setState((state) => {
-          return { ...state, [name]: checkTitle(value, state.risk) };
-        });
-        break;
-      default:
-        setState((state) => {
-          return { ...state, [name]: value };
-        });
-        break;
-    }
-  }
-};
-
 export const checkQaOrGasFinding = (risk) => {
-  return risk.slice(0, 1) === "G" || risk.slice(0, 1) === "Q";
+  if (risk === ""){
+    return true;
+  } else {
+    return risk.slice(0, 1) === "G" || risk.slice(0, 1) === "Q";
+  }
 };
